requirements.txt
Flask==2.1.2
requests==2.27.1
python-dotenv==1.0.1
Werkzeug==2.2.2
psycopg2-binary==2.9.9
pandas==2.1.4
sqlalchemy==2.0.23